refactor(azure-devops): extract reusable ADO/TimeLog config lib from CLI commands Extract core config logic into lib/ so both CLI and Clarity UI can use it without duplication: - lib/ado-configure.ts: checkAzureCliLogin(), buildAdoConfig(), saveAdoConfig() - lib/timelog-configure.ts: fetchTimeLogFunctionsKey() with proper JSDoc CLI commands now import from lib/ instead of inlining the logic. Uses AZURE_DEVOPS_RESOURCE_ID constant instead of hardcoded GUID.
